fix: Vercelデプロイ用の設定修正
- src/index.tsからgetApp()関数をexport
- api/serverless.tsを追加（参考: github-webhook-bridge）
- vercel.jsonをapi/serverless.tsを指すよう修正
