+++ src/index.ts
@@ -1,11 +1,10 @@
 import fastify from 'fastify'
 import cors from '@fastify/cors'
-import { IncomingMessage, ServerResponse } from 'node:http'
 import { loadConfig } from './config.js'
 import { RSSProcessor } from './rss-processor.js'
 import { TranslateRequest, TranslateResponse } from './types.js'
 
-async function start() {
+export async function getApp() {
   const config = loadConfig()
   const app = fastify({
     logger: true,
@@ -67,6 +66,13 @@ async function start() {
     }
   })
 
+  return app
+}
+
+async function start() {
+  const config = loadConfig()
+  const app = await getApp()
+
   // Start server
   try {
     await app.listen({
@@ -78,67 +84,6 @@ async function start() {
     app.log.error(error)
     throw new Error('Failed to start server')
   }
-}
-
-// Vercel serverless function export
-export default async function handler(
-  req: IncomingMessage,
-  res: ServerResponse
-) {
-  const config = loadConfig()
-  const app = fastify({
-    logger: false,
-  })
-
-  await app.register(cors, {
-    origin: true,
-  })
-
-  const rssProcessor = new RSSProcessor(config.gasUrl)
-
-  app.get<{
-    Querystring: TranslateRequest
-  }>('/', async (request, reply) => {
-    const { url, sourceLang, targetLang } = request.query
-
-    if (!url) {
-      const response: TranslateResponse = {
-        status: 'error',
-        error: 'URL parameter is required',
-      }
-      return reply.code(400).send(response)
-    }
-
-    try {
-      const translatedRSS = await rssProcessor.processRSSFeed(
-        url,
-        sourceLang ?? config.defaultSourceLang ?? 'auto',
-        targetLang ?? config.defaultTargetLang ?? 'ja'
-      )
-
-      if (!translatedRSS) {
-        const response: TranslateResponse = {
-          status: 'error',
-          error: 'Failed to process RSS feed',
-        }
-        return await reply.code(500).send(response)
-      }
-
-      return await reply
-        .code(200)
-        .header('Content-Type', 'application/rss+xml; charset=utf-8')
-        .send(translatedRSS)
-    } catch {
-      const response: TranslateResponse = {
-        status: 'error',
-        error: 'Internal server error',
-      }
-      return await reply.code(500).send(response)
-    }
-  })
-
-  await app.ready()
-  app.server.emit('request', req, res)
 }
 
 // Run if called directly
